# Optimize the performance of character comparison