change the char comparison of Email, and add the benchmark:

Change before:
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	2000000000	         0.34 ns/op
PASS
ok  	github.com/MonkSunBoy/is	19.904s
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	2000000000	         0.33 ns/op
PASS
ok  	github.com/MonkSunBoy/is	19.276s
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	2000000000	         0.33 ns/op
PASS
ok  	github.com/MonkSunBoy/is	20.348s

Change after:
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	2000000000	         0.12 ns/op
PASS
ok  	github.com/MonkSunBoy/is	3.925s
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	2000000000	         0.13 ns/op
PASS
ok  	github.com/MonkSunBoy/is	3.940s
monkboy@monkboy:~/work/go/src/github.com/MonkSunBoy/is$ go test -bench .
BenchmarkEmail-4   	1000000000	         0.26 ns/op
PASS
ok  	github.com/MonkSunBoy/is	3.402s